backendAssignment/requirements.txt
Django==2.1.1
django-timezone-field==4.0
djangorestframework==3.11.0
mysqlclient==1.4.6
pytz==2020.1

